requirements.txt
mysql-connector-python==8.0.30
pkg_resources==0.0.0
protobuf==3.20.1
PyYAML==6.0
